# api/openapi/loco-place-v1.yaml
openapi: 3.1.0
info:
  title: Loco Place Public API
  version: 1.0.0-rc.1
  description: |
    Channel-neutral public contract for web, Telegram Mini App and native clients.
    Unknown enum values must be preserved by generated clients as UNKNOWN rather than causing decoding failure.
servers:
  - url: https://loco-place.com
...
paths:
...
  /api/v1/auth/tma/session:
    post:
      tags: [Authentication]
      operationId: createTmaSession
      security: [{ tmaBootstrap: [] }]
      responses:
        '204': { description: Cookie session created }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/auth/native/challenges:
    post:
      tags: [Authentication]
      operationId: createNativeChallenge
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateNativeChallengeRequest' }
      responses:
        '200':
          description: Nonce-bound login challenge
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NativeChallenge' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/auth/native/google:
    post:
      tags: [Authentication]
      operationId: loginWithGoogle
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NativeProviderLoginRequest' }
      responses:
        '200':
          description: First-party native credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionTokens' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/native/apple:
    post:
      tags: [Authentication]
      operationId: loginWithApple
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NativeProviderLoginRequest' }
      responses:
        '200':
          description: First-party native credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionTokens' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/v1/auth/native/telegram/attempts/{attemptId}/exchange:
    post:
      tags: [Authentication]
      operationId: exchangeTelegramLoginAttempt
      security: []
      parameters:
        - name: attemptId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verifier, clientType]
              properties:
                verifier: { type: string, minLength: 1 }
                clientType: { type: string, enum: [IOS, ANDROID, UNKNOWN] }
      responses:
        '200':
          description: First-party native credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionTokens' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/sessions/refresh:
    post:
      tags: [Authentication]
      operationId: refreshSession
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string, pattern: '^lp_rt_' }
      responses:
        '200':
          description: Rotated credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SessionTokens' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/auth/sessions/current:
    delete:
      tags: [Authentication]
      operationId: logoutCurrentSession
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      responses:
        '204': { description: Current session revoked }
  /api/v1/auth/sessions:
    delete:
      tags: [Authentication]
      operationId: revokeAllSessions
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      responses:
        '204': { description: Every customer session revoked }
...
  /api/v1/account/notifications:
    get:
      tags: [Account]
      operationId: getNotifications
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      parameters:
        - name: cursor
          in: query
          required: false
          schema: { type: string }
        - name: size
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: Stable newest-first notification page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationCursorPage' }
  /api/v1/account/notifications/unread-count:
    get:
      tags: [Account]
      operationId: getNotificationUnreadCount
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      responses:
        '200':
          description: Current unread notification count
          content:
            application/json:
              schema: { $ref: '#/components/schemas/NotificationUnreadCount' }
  /api/v1/account/notifications/{notificationId}/read:
    post:
      tags: [Account]
      operationId: markNotificationRead
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      parameters:
        - name: notificationId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: Notification marked read }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/account/notifications/read-all:
    post:
      tags: [Account]
      operationId: markAllNotificationsRead
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      responses:
        '204': { description: Every customer notification marked read }
...
  /api/v1/transfer/bookings:
    get:
      tags: [Transfer]
      operationId: getTransferBookings
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Stable newest-first Transfer booking page
          content: { application/json: { schema: { $ref: '#/components/schemas/TransferBookingCursorPage' } } }
    post:
      tags: [Transfer]
      operationId: createTransferBooking
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateTransferBookingRequest' } } }
      responses:
        '201':
          description: Transfer booking created
          content: { application/json: { schema: { $ref: '#/components/schemas/TransferBooking' } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/transfer/configuration:
    get:
      tags: [Transfer]
      operationId: getTransferConfiguration
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      responses:
        '200':
          description: Transfer airports, vehicles, prices and date policy
          content: { application/json: { schema: { $ref: '#/components/schemas/TransferConfiguration' } } }
  /api/v1/transfer/quote:
    post:
      tags: [Transfer]
      operationId: quoteTransfer
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TransferQuoteRequest' } } }
      responses:
        '200':
          description: Authoritative Transfer quote
          content: { application/json: { schema: { $ref: '#/components/schemas/TransferQuote' } } }
  /api/v1/transfer/bookings/{bookingId}:
    get:
      tags: [Transfer]
      operationId: getTransferBooking
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      parameters: [{ $ref: '#/components/parameters/BookingId' }]
      responses:
        '200':
          description: Customer-owned Transfer booking
          content: { application/json: { schema: { $ref: '#/components/schemas/TransferBooking' } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/transfer/bookings/{bookingId}/repeat-shown:
    post:
      tags: [Transfer]
      operationId: recordTransferRepeatShown
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      parameters: [{ $ref: '#/components/parameters/BookingId' }]
      responses: { '204': { description: Repeat opportunity impression recorded } }
  /api/v1/transfer/bookings/{bookingId}/repeat-prefill:
    post:
      tags: [Transfer]
      operationId: getTransferRepeatPrefill
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      parameters: [{ $ref: '#/components/parameters/BookingId' }]
      responses:
        '200':
          description: Verified repeat Transfer context
          content: { application/json: { schema: { $ref: '#/components/schemas/TransferRepeatPrefill' } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/transfer/bookings/{bookingId}/cancel:
    post:
      tags: [Transfer]
      operationId: cancelTransferBooking
      security: [{ bearerAuth: [] }, { cookieSession: [] }]
      parameters: [{ $ref: '#/components/parameters/BookingId' }]
      responses:
        '200':
          description: Cancelled Transfer booking
          content: { application/json: { schema: { $ref: '#/components/schemas/TransferBooking' } } }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 200 }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
    PageSize:
      name: size
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
...
    BookingId:
      name: bookingId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: First-party opaque access token. Provider credentials are never accepted here.
    cookieSession:
      type: apiKey
      in: cookie
      name: SESSION
      description: Browser/TMA HttpOnly session. Mutations also require X-XSRF-TOKEN.
    tmaBootstrap:
      type: apiKey
      in: header
      name: Authorization
      description: Authorization header uses `tma <initData>` only for the TMA session exchange.
  responses:
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Unauthorized:
      description: Authentication failed
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Conflict:
      description: Business or security conflict
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
...
  schemas:
...
    IdentityProvider:
      type: string
      enum: [GOOGLE, APPLE, TELEGRAM, UNKNOWN]
...
    CreateNativeChallengeRequest:
      type: object
      required: [provider, clientType]
      properties:
        provider: { $ref: '#/components/schemas/IdentityProvider' }
        clientType: { type: string, enum: [IOS, ANDROID, UNKNOWN] }
    NativeChallenge:
      type: object
      required: [challengeId, provider, nonce, expiresAt]
      properties:
        challengeId: { type: string, format: uuid }
        provider: { $ref: '#/components/schemas/IdentityProvider' }
        nonce: { type: string }
        expiresAt: { type: string, format: date-time }
    NativeProviderLoginRequest:
      type: object
      required: [challengeId, clientType, identityToken]
      properties:
        challengeId: { type: string, format: uuid }
        clientType: { type: string, enum: [IOS, ANDROID, UNKNOWN] }
        identityToken: { type: string }
        authorizationCode: { type: [string, 'null'] }
    SessionTokens:
      type: object
      required: [sessionId, tokenType, accessToken, refreshToken, accessExpiresAt, refreshExpiresAt, absoluteExpiresAt]
      properties:
        sessionId: { type: string, format: uuid }
        tokenType: { type: string, const: Bearer }
        accessToken: { type: string, pattern: '^lp_at_' }
        refreshToken: { type: string, pattern: '^lp_rt_' }
        accessExpiresAt: { type: string, format: date-time }
        refreshExpiresAt: { type: string, format: date-time }
        absoluteExpiresAt: { type: string, format: date-time }
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: string, pattern: '^-?[0-9]+(?:\.[0-9]{1,2})?$' }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
...
    RentalTransferSourceRequest:
      type: object
      required: [bookingId, context]
      properties:
        bookingId: { type: integer, format: int64, minimum: 1 }
        context: { type: string, enum: [ARRIVAL, CHECKOUT, UNKNOWN] }
    CreateTransferBookingRequest:
      type: object
      required: [direction, airportId, vehicleTypeId, pickupDate, pickupTime, address,
                 passengerCount, luggageCount, phone]
      properties:
        direction: { type: string, enum: [TO_AIRPORT, FROM_AIRPORT, UNKNOWN] }
        airportId: { type: integer, format: int64, minimum: 1 }
        vehicleTypeId: { type: integer, format: int64, minimum: 1 }
        pickupDate: { type: string, format: date }
        pickupTime: { type: string, format: time }
        address: { type: string, minLength: 1, maxLength: 1000 }
        passengerCount: { type: integer, minimum: 1, maximum: 50 }
        luggageCount: { type: integer, minimum: 0, maximum: 50 }
        flightNumber: { type: [string, 'null'], maxLength: 64 }
        scheduledArrivalTime: { type: [string, 'null'], format: time }
        phone: { type: string, minLength: 1, maxLength: 40 }
        comment: { type: [string, 'null'], maxLength: 1000 }
        repeatFromBookingId: { type: [integer, 'null'], format: int64, minimum: 1 }
        rentalSource:
          oneOf:
            - $ref: '#/components/schemas/RentalTransferSourceRequest'
            - type: 'null'
        benefit: { type: [string, 'null'], enum: [RENTAL_FIRST_TRANSFER, UNKNOWN, null] }
    TransferQuoteRequest:
      type: object
      required: [direction, airportId, vehicleTypeId]
      properties:
        direction: { type: string, enum: [TO_AIRPORT, FROM_AIRPORT, UNKNOWN] }
        airportId: { type: integer, format: int64, minimum: 1 }
        vehicleTypeId: { type: integer, format: int64, minimum: 1 }
        rentalSource:
          oneOf:
            - $ref: '#/components/schemas/RentalTransferSourceRequest'
            - type: 'null'
        benefit: { type: [string, 'null'], enum: [RENTAL_FIRST_TRANSFER, UNKNOWN, null] }
    TransferQuote:
      type: object
      required: [baseAmount, discountAmount, payableAmount, currency, appliedBenefit]
      properties:
        baseAmount: { type: number }
        discountAmount: { type: number }
        payableAmount: { type: number }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        appliedBenefit: { type: [string, 'null'], enum: [RENTAL_FIRST_TRANSFER, UNKNOWN, null] }
    TransferAirport:
      type: object
      required: [id, code, nameRu, nameEn]
      properties:
        id: { type: integer, format: int64 }
        code: { type: string }
        nameRu: { type: string }
        nameEn: { type: string }
    TransferVehicleType:
      type: object
      required: [id, code, nameRu, nameEn, maxPassengers, maxLuggage]
      properties:
        id: { type: integer, format: int64 }
        code: { type: string }
        nameRu: { type: string }
        nameEn: { type: string }
        maxPassengers: { type: integer }
        maxLuggage: { type: integer }
    TransferPrice:
      type: object
      required: [airportId, vehicleTypeId, direction, amount, currency]
      properties:
        airportId: { type: integer, format: int64 }
        vehicleTypeId: { type: integer, format: int64 }
        direction: { type: string, enum: [TO_AIRPORT, FROM_AIRPORT, UNKNOWN] }
        amount: { type: number }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
    TransferConfiguration:
      type: object
      required: [earliestBookingDate, latestBookingDate, timeSlotMinutes, airports, vehicleTypes, prices]
      properties:
        earliestBookingDate: { type: string, format: date }
        latestBookingDate: { type: string, format: date }
        timeSlotMinutes: { type: integer, minimum: 1 }
        airports:
          type: array
          items: { $ref: '#/components/schemas/TransferAirport' }
        vehicleTypes:
          type: array
          items: { $ref: '#/components/schemas/TransferVehicleType' }
        prices:
          type: array
          items: { $ref: '#/components/schemas/TransferPrice' }
    TransferBooking:
      type: object
      required: [id, direction, airportCode, airportNameRu, airportNameEn, vehicleCode,
                 vehicleNameRu, vehicleNameEn, pickupDate, pickupTime, address, passengerCount,
                 luggageCount, flightNumber, scheduledArrivalTime, customerName, phone, comment,
                 basePriceAmount, discountAmount, priceAmount, priceCurrency, appliedBenefit,
                 benefitRate, status, driverId, driverName, createdAt, confirmedAt, completedAt,
                 cancelledAt, rejectedAt, statusReason, money]
      properties:
        id: { type: integer, format: int64 }
        direction: { type: string, enum: [TO_AIRPORT, FROM_AIRPORT, UNKNOWN] }
        airportCode: { type: string }
        airportNameRu: { type: string }
        airportNameEn: { type: string }
        vehicleCode: { type: string }
        vehicleNameRu: { type: string }
        vehicleNameEn: { type: string }
        pickupDate: { type: string, format: date }
        pickupTime: { type: string, format: time }
        address: { type: string }
        passengerCount: { type: integer }
        luggageCount: { type: integer }
        flightNumber: { type: [string, 'null'] }
        scheduledArrivalTime: { type: [string, 'null'], format: time }
        customerName: { type: string }
        phone: { type: string }
        comment: { type: [string, 'null'] }
        basePriceAmount: { type: number }
        discountAmount: { type: number }
        priceAmount: { type: number }
        priceCurrency: { type: string, pattern: '^[A-Z]{3}$' }
        appliedBenefit: { type: [string, 'null'], enum: [RENTAL_FIRST_TRANSFER, UNKNOWN, null] }
        benefitRate: { type: [number, 'null'] }
        status: { type: string, enum: [REQUESTED, CONFIRMED, COMPLETED, CANCELLED, REJECTED, UNKNOWN] }
        driverId: { type: [integer, 'null'], format: int64 }
        driverName: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        confirmedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        cancelledAt: { type: [string, 'null'], format: date-time }
        rejectedAt: { type: [string, 'null'], format: date-time }
        statusReason: { type: [string, 'null'] }
        money: { $ref: '#/components/schemas/Money' }
    TransferBookingCursorPage:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/TransferBooking' }
        nextCursor: { type: [string, 'null'] }
        hasMore: { type: boolean }
    TransferRepeatPrefill:
      type: object
      required: [sourceBookingId, direction, airportId, vehicleTypeId, address, passengerCount, luggageCount]
      properties:
        sourceBookingId: { type: integer, format: int64 }
        direction: { type: string, enum: [TO_AIRPORT, FROM_AIRPORT, UNKNOWN] }
        airportId: { type: [integer, 'null'], format: int64 }
        vehicleTypeId: { type: [integer, 'null'], format: int64 }
        address: { type: string }
        passengerCount: { type: integer }
        luggageCount: { type: integer }
...
    ActionTarget:
      oneOf:
        - $ref: '#/components/schemas/OpenTransactionAction'
        - $ref: '#/components/schemas/OpenCleaningHistoryAction'
        - $ref: '#/components/schemas/RepeatCleaningAction'
        - $ref: '#/components/schemas/RepeatTransferAction'
        - $ref: '#/components/schemas/StartRentalCleaningAction'
        - $ref: '#/components/schemas/StartRentalTransferAction'
        - $ref: '#/components/schemas/OpenAdminTransactionAction'
        - $ref: '#/components/schemas/OpenSupportCaseAction'
      discriminator:
        propertyName: type
        mapping:
          OPEN_TRANSACTION: '#/components/schemas/OpenTransactionAction'
          OPEN_CLEANING_HISTORY: '#/components/schemas/OpenCleaningHistoryAction'
          REPEAT_CLEANING: '#/components/schemas/RepeatCleaningAction'
          REPEAT_TRANSFER: '#/components/schemas/RepeatTransferAction'
          START_RENTAL_CLEANING: '#/components/schemas/StartRentalCleaningAction'
          START_RENTAL_TRANSFER: '#/components/schemas/StartRentalTransferAction'
          OPEN_ADMIN_TRANSACTION: '#/components/schemas/OpenAdminTransactionAction'
          OPEN_SUPPORT_CASE: '#/components/schemas/OpenSupportCaseAction'
    OpenTransactionAction:
      type: object
      required: [type, service, entityId]
      properties:
        type: { type: string, const: OPEN_TRANSACTION }
        service: { type: string, enum: [CLEANING, RENTAL, TRANSFER, UNKNOWN] }
        entityId: { type: integer, format: int64 }
    OpenCleaningHistoryAction:
      type: object
      required: [type]
      properties: { type: { type: string, const: OPEN_CLEANING_HISTORY } }
    RepeatCleaningAction:
      type: object
      required: [type, sourceOrderId]
      properties:
        type: { type: string, const: REPEAT_CLEANING }
        sourceOrderId: { type: integer, format: int64 }
    RepeatTransferAction:
      type: object
      required: [type, sourceBookingId]
      properties:
        type: { type: string, const: REPEAT_TRANSFER }
        sourceBookingId: { type: integer, format: int64 }
    StartRentalCleaningAction:
      type: object
      required: [type, rentalBookingId]
      properties:
        type: { type: string, const: START_RENTAL_CLEANING }
        rentalBookingId: { type: integer, format: int64 }
    StartRentalTransferAction:
      type: object
      required: [type, rentalBookingId, context]
      properties:
        type: { type: string, const: START_RENTAL_TRANSFER }
        rentalBookingId: { type: integer, format: int64 }
        context: { type: string, enum: [ARRIVAL, CHECKOUT, UNKNOWN] }
    OpenAdminTransactionAction:
      type: object
      required: [type, service, entityId]
      properties:
        type: { type: string, const: OPEN_ADMIN_TRANSACTION }
        service: { type: string, enum: [CLEANING, RENTAL, TRANSFER, UNKNOWN] }
        entityId: { type: integer, format: int64 }
    OpenSupportCaseAction:
      type: object
      required: [type, caseId]
      properties:
        type: { type: string, const: OPEN_SUPPORT_CASE }
        caseId: { type: integer, format: int64 }
    CustomerNotification:
      type: object
      required: [id, type, action, createdAt]
      properties:
        id: { type: integer, format: int64 }
        type: { type: string }
        action: { $ref: '#/components/schemas/ActionTarget' }
        createdAt: { type: string, format: date-time }
        readAt: { type: [string, 'null'], format: date-time }
...
    NotificationUnreadCount:
      type: object
      required: [unreadCount]
      properties:
        unreadCount: { type: integer, minimum: 0 }
    NotificationCursorPage:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CustomerNotification' }
        nextCursor: { type: [string, 'null'] }
        hasMore: { type: boolean }
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }
        requestId: { type: string }